backend/requirements.txt
flask
flask-socketio
web3.py
requests
langchain
openai
coinbase-agentkit
python-dotenv